# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Stores information about dashboards and their widgets.",
        "title": "Dashboard API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/dashboards": {
            "get": {
                "description": "Returns all dashboards for the current user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "List dashboards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.Dashboard"
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new dashboard for the current user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Create dashboard",
                "parameters": [
                    {
                        "description": "Dashboard payload",
                        "name": "dashboard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Dashboard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Dashboard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dashboards/{id}": {
            "get": {
                "description": "Returns a dashboard by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Get dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Dashboard"
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates dashboard metadata by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Update dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dashboard payload",
                        "name": "dashboard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Dashboard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Dashboard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a dashboard by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Delete dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/doc": {
            "get": {
                "description": "Returns the generated Swagger document for this service.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documentation"
                ],
                "summary": "Get OpenAPI document",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/widgets/name/{dashboardId}/{widgetId}": {
            "patch": {
                "description": "Updates the name of a widget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widgets"
                ],
                "summary": "Update widget name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "dashboardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Widget ID",
                        "name": "widgetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New widget name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/widgets/positions": {
            "patch": {
                "description": "Updates positions for multiple widgets.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widgets"
                ],
                "summary": "Update widget positions",
                "parameters": [
                    {
                        "description": "Widget position updates",
                        "name": "positions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.WidgetPosition"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/widgets/properties/{dashboardId}/{widgetId}": {
            "patch": {
                "description": "Updates the complete properties object of a widget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widgets"
                ],
                "summary": "Update widget properties",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "dashboardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Widget ID",
                        "name": "widgetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New properties object",
                        "name": "properties",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/widgets/properties/{property}/{dashboardId}/{widgetId}": {
            "patch": {
                "description": "Updates a single widget property by property key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widgets"
                ],
                "summary": "Update one widget property",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property name",
                        "name": "property",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "dashboardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Widget ID",
                        "name": "widgetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New property value",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/widgets/{dashboardId}": {
            "post": {
                "description": "Creates a new widget in a dashboard.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widgets"
                ],
                "summary": "Create widget",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard ID",
                        "name": "dashboardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Widget payload",
                        "name": "widget",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Widget"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Widget"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "lib.Dashboard": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "refresh_time": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Widget"
                    }
                }
            }
        },
        "lib.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "lib.Widget": {
            "type": "object",
            "properties": {
                "h": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "properties": {},
                "type": {
                    "type": "string"
                },
                "w": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "lib.WidgetPosition": {
            "type": "object",
            "properties": {
                "dashboardDestination": {
                    "type": "string"
                },
                "dashboardOrigin": {
                    "type": "string"
                },
                "h": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "w": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}